Adjust and refactor all `MaterialButton` tests into its respective file (#43528)
* Adjust and refactor all MaterialButton tests into its respective file
